Move get_X_y to conftest fixture
